+++ src/utils/introspection/test/augmentations.d.ts
@@ -1,4 +1,5 @@
 import type { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
+import { Contract } from '@ethersproject/contracts';
 
 // augmentations.d.ts
 declare module 'mocha' {
@@ -7,5 +8,7 @@ declare module 'mocha' {
         addr1: SignerWithAddress;
         addr2: SignerWithAddress;
         addr3: SignerWithAddress;
+
+        contractUnderTest: Contract;
     }
 }