+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 As of 12/13/2022, this repo has been renamed from "nftc-open-contracts" to "nftc-contracts". Please start using package version 1.3.0 or later.
 
 
+## Version -- 1.4.7
+- SupportsInterface introspection tests now validate that appropriate functions exist to match ERC721 spec.
+- Implemented basic ERC165 checks for 721OZ, 721A and 721SolBase
+
 
 ## Version -- 1.4.6 (Not published)
 - Finish first pass at getting library hooks to work.